test: extract provider sections to tests/providers/ (batch B, #1440)
