+++ test-all.mjs
@@ -4790,198 +4790,6 @@ try {
   fail(`CJK rendering test crashed: ${e.message}`);
 }
 
-console.log('\n24. Provider — ibm');
-
-try {
-  const ibm = (await import(pathToFileURL(join(ROOT, 'providers/ibm.mjs')).href)).default;
-  const { parseIbmResponse, buildPostFilter } = await import(pathToFileURL(join(ROOT, 'providers/ibm.mjs')).href);
-
-  if (ibm.id === 'ibm') pass('ibm.id is "ibm"');
-  else fail(`ibm.id is ${JSON.stringify(ibm.id)}`);
-
-  // buildPostFilter — empty config yields no filter terms
-  if (buildPostFilter({}).bool.must.length === 0) pass('buildPostFilter({}) → no must terms');
-  else fail(`buildPostFilter({}) = ${JSON.stringify(buildPostFilter({}))}`);
-
-  // buildPostFilter — country + categories produce the expected facet terms
-  const pf = buildPostFilter({ country: 'Germany', categories: ['Software Engineering', 'Data & Analytics'] });
-  const countryTerm = pf.bool.must.find(m => m.term && m.term.field_keyword_05);
-  const catTerm = pf.bool.must.find(m => m.bool && m.bool.should);
-  if (countryTerm?.term.field_keyword_05 === 'Germany' && catTerm?.bool.should.length === 2) {
-    pass('buildPostFilter maps country → field_keyword_05 and categories → field_keyword_08 should[]');
-  } else {
-    fail(`buildPostFilter facets = ${JSON.stringify(pf)}`);
-  }
-
-  // buildPostFilter — sanitizes empty/non-string category entries
-  const sanitized = buildPostFilter({ categories: ['Valid', '', '   ', 42, null] });
-  const sanitizedShould = sanitized.bool.must.find(m => m.bool && m.bool.should)?.bool.should;
-  if (sanitizedShould?.length === 1 && sanitizedShould[0].term.field_keyword_08 === 'Valid') {
-    pass('buildPostFilter drops empty/non-string category entries');
-  } else {
-    fail(`buildPostFilter sanitization = ${JSON.stringify(sanitizedShould)}`);
-  }
-
-  // parseIbmResponse — happy path, location assembled from keyword_19 · keyword_17
-  const sample = {
-    hits: {
-      hits: [
-        { _source: { title: 'ML Engineer', url: 'https://ibm.com/careers/1', field_keyword_19: 'Berlin, Germany', field_keyword_17: 'Hybrid' } },
-        { _source: { title: 'Data Scientist', url: 'https://ibm.com/careers/2', field_keyword_19: 'Remote' } },
-      ],
-    },
-  };
-  const jobs = parseIbmResponse(sample);
-  if (jobs.length === 2 && jobs[0].company === 'IBM') pass('parseIbmResponse extracts 2 jobs with company "IBM"');
-  else fail(`parseIbmResponse returned ${JSON.stringify(jobs)}`);
-
-  if (jobs[0].location === 'Berlin, Germany · Hybrid') pass('parseIbmResponse joins location · work mode');
-  else fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}`);
-
-  if (jobs[1].location === 'Remote') pass('parseIbmResponse omits the separator when work mode is absent');
-  else fail(`row 1 location = ${JSON.stringify(jobs[1]?.location)}`);
-
-  // parseIbmResponse — drops title-less, url-less, and non-http(s) entries
-  const dirty = parseIbmResponse({
-    hits: {
-      hits: [
-        { _source: { title: '', url: 'https://ibm.com/careers/3' } },
-        { _source: { title: 'No URL' } },
-        { _source: { title: 'Bad scheme', url: 'ftp://ibm.com/careers/4' } },
-        { _source: { title: 'Good', url: 'https://ibm.com/careers/5' } },
-      ],
-    },
-  });
-  if (dirty.length === 1 && dirty[0].title === 'Good') pass('parseIbmResponse drops title-less, url-less, and non-http(s) entries');
-  else fail(`parseIbmResponse dirty = ${JSON.stringify(dirty)}`);
-
-  // parseIbmResponse — throws on unexpected shape (endpoint drift surfaces loudly)
-  let drifted = false;
-  try { parseIbmResponse({ results: [] }); } catch { drifted = true; }
-  if (drifted) pass('parseIbmResponse throws when hits.hits[] is missing');
-  else fail('parseIbmResponse should throw on unexpected API response shape');
-
-  // fetch() — paginates until a short page, via mock ctx
-  let calls = 0;
-  const mockCtx = {
-    fetchJson: async (url, opts) => {
-      calls++;
-      if (url !== 'https://www-api.ibm.com/search/api/v2') throw new Error(`unexpected url ${url}`);
-      if (opts?.method !== 'POST') throw new Error('Expected POST');
-      // Page 1: a full page (30 hits) → keep paging; page 2: short page → stop.
-      const n = calls === 1 ? 30 : 2;
-      const hits = Array.from({ length: n }, (_, i) => ({
-        _source: { title: `Role ${calls}-${i}`, url: `https://ibm.com/careers/${calls}-${i}` },
-      }));
-      return { hits: { hits } };
-    },
-  };
-  const fetched = await ibm.fetch({ name: 'IBM', ibm: { country: 'Germany' } }, mockCtx);
-  if (calls === 2 && fetched.length === 32) pass('ibm.fetch() paginates and stops on the first short page');
-  else fail(`ibm.fetch() made ${calls} calls, returned ${fetched.length} jobs`);
-
-} catch (e) {
-  fail(`ibm provider tests crashed: ${e.message}`);
-}
-
-console.log('\n26. Provider — bamboohr');
-try {
-  const bamboohr = (await import(pathToFileURL(join(ROOT, 'providers/bamboohr.mjs')).href)).default;
-  const { parseBambooHRResponse } = await import(pathToFileURL(join(ROOT, 'providers/bamboohr.mjs')).href);
-
-  if (bamboohr.id === 'bamboohr') pass('bamboohr.id is "bamboohr"');
-  else fail(`bamboohr.id is ${JSON.stringify(bamboohr.id)}`);
-
-  // detect: <tenant>.bamboohr.com → /careers/list
-  const hit = bamboohr.detect({ name: 'Acme', careers_url: 'https://acme.bamboohr.com/careers' });
-  if (hit && hit.url === 'https://acme.bamboohr.com/careers/list') {
-    pass('bamboohr.detect() resolves <tenant>.bamboohr.com → /careers/list');
-  } else {
-    fail(`bamboohr.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect: honours an explicit api: URL
-  const apiHit = bamboohr.detect({ name: 'Acme', api: 'https://acme.bamboohr.com' });
-  if (apiHit && apiHit.url === 'https://acme.bamboohr.com/careers/list') pass('bamboohr.detect() honours explicit api: URL');
-  else fail(`bamboohr.detect() api: returned ${JSON.stringify(apiHit)}`);
-
-  // detect: null for non-bamboohr URLs
-  if (bamboohr.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('bamboohr.detect() returns null for non-bamboohr URLs');
-  } else {
-    fail('bamboohr.detect() should return null for non-bamboohr URLs');
-  }
-
-  // detect: null for non-string careers_url
-  if (bamboohr.detect({ name: 'X', careers_url: null }) === null && bamboohr.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('bamboohr.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('bamboohr.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: bamboohr.com in the PATH (not host) must not be detected.
-  if (bamboohr.detect({ name: 'Spoof', careers_url: 'https://evil.example/acme.bamboohr.com/foo' }) === null) {
-    pass('bamboohr.detect() rejects path-spoofed URLs');
-  } else {
-    fail('bamboohr.detect() must reject path-spoofed URLs');
-  }
-
-  // parseBambooHRResponse — real BambooHR list shape
-  const sample = {
-    meta: {},
-    result: [
-      { id: '15', jobOpeningName: 'IT Security Engineer', location: { city: 'Mayfair', state: 'London, City of' }, isRemote: null },
-      { id: 22, jobOpeningName: 'Android Engineer', location: { city: 'Bengaluru', state: 'Karnataka' }, isRemote: 1 },
-      { id: '7', jobOpeningName: '', location: { city: 'X' } },          // no title → dropped
-      { jobOpeningName: 'No ID Role', location: { city: 'Y' } },          // no id → dropped
-      { id: '   ', jobOpeningName: 'Blank ID Role', location: { city: 'Z' } }, // blank/whitespace id → dropped
-    ],
-  };
-  const jobs = parseBambooHRResponse(sample, 'Acme', 'https://acme.bamboohr.com');
-  if (jobs.length === 2) pass('parseBambooHRResponse keeps rows with non-empty id + title, drops the rest');
-  else fail(`parseBambooHRResponse returned ${jobs.length} jobs (expected 2)`);
-
-  if (!jobs.some(j => j.title === 'Blank ID Role')) pass('parseBambooHRResponse drops blank/whitespace-id rows (no /careers/ URL)');
-  else fail('parseBambooHRResponse should drop blank/whitespace-id rows');
-
-  if (jobs[0]?.url === 'https://acme.bamboohr.com/careers/15' && jobs[0]?.company === 'Acme') {
-    pass('parseBambooHRResponse builds <origin>/careers/<id> URL');
-  } else {
-    fail(`parseBambooHRResponse url was ${jobs[0]?.url}`);
-  }
-
-  if (jobs[0]?.location === 'Mayfair, London, City of') pass('parseBambooHRResponse joins city + state');
-  else fail(`parseBambooHRResponse location[0] was ${JSON.stringify(jobs[0]?.location)}`);
-
-  if (jobs[1]?.location === 'Bengaluru, Karnataka, Remote') pass('parseBambooHRResponse appends Remote when isRemote is set');
-  else fail(`parseBambooHRResponse location[1] was ${JSON.stringify(jobs[1]?.location)}`);
-
-  if (jobs[1]?.url === 'https://acme.bamboohr.com/careers/22') pass('parseBambooHRResponse coerces numeric id to URL');
-  else fail(`parseBambooHRResponse numeric-id url was ${jobs[1]?.url}`);
-
-  // empty / malformed payloads → []
-  if (parseBambooHRResponse({}, 'X', 'https://x.bamboohr.com').length === 0) pass('parseBambooHRResponse empty {} → []');
-  else fail('parseBambooHRResponse should return [] for {}');
-  if (parseBambooHRResponse({ result: null }, 'X', 'https://x.bamboohr.com').length === 0) pass('parseBambooHRResponse result:null → []');
-  else fail('parseBambooHRResponse should return [] for result:null');
-
-  // fetch() — via mock ctx, asserts the resolved URL + SSRF redirect pinning + parsing
-  let fetchedUrl = '';
-  let fetchedOpts;
-  const mockCtx = {
-    fetchJson: async (url, opts) => { fetchedUrl = url; fetchedOpts = opts; return sample; },
-  };
-  const fetched = await bamboohr.fetch({ name: 'Acme', careers_url: 'https://acme.bamboohr.com/careers' }, mockCtx);
-  if (fetchedUrl === 'https://acme.bamboohr.com/careers/list' && fetchedOpts?.redirect === 'error' && fetched.length === 2) {
-    pass('bamboohr.fetch() calls /careers/list with redirect:error and returns parsed jobs');
-  } else {
-    fail(`bamboohr.fetch() url=${fetchedUrl} redirect=${JSON.stringify(fetchedOpts)} jobs=${fetched.length}`);
-  }
-
-} catch (e) {
-  fail(`bamboohr provider tests crashed: ${e.message}`);
-}
-
 // ── 27. ATS LIGATURE SUPPRESSION ────────────────────────────────
 
 console.log('\n27. ATS ligature suppression');
@@ -5013,111 +4821,6 @@ try {
   }
 } catch (e) {
   fail(`ATS ligature suppression test crashed: ${e.message}`);
-}
-
-// ── 27. Provider — breezy ───────────────────────────────────────
-console.log('\n27. Provider — breezy');
-
-try {
-  const breezy = (await import(pathToFileURL(join(ROOT, 'providers/breezy.mjs')).href)).default;
-  const { parseBreezyResponse } = await import(pathToFileURL(join(ROOT, 'providers/breezy.mjs')).href);
-
-  if (breezy.id === 'breezy') pass('breezy.id is "breezy"');
-  else fail(`breezy.id is ${JSON.stringify(breezy.id)}`);
-
-  // detect: careers_url with a path still resolves the tenant /json feed
-  const hit = breezy.detect({ name: 'New Incentives', careers_url: 'https://new-incentives.breezy.hr/' });
-  if (hit && hit.url === 'https://new-incentives.breezy.hr/json') {
-    pass('breezy.detect() resolves <tenant>.breezy.hr → /json feed');
-  } else {
-    fail(`breezy.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect: explicit api: URL is honoured over careers_url
-  const apiHit = breezy.detect({ name: 'X', api: 'https://acme.breezy.hr', careers_url: 'https://example.com' });
-  if (apiHit && apiHit.url === 'https://acme.breezy.hr/json') {
-    pass('breezy.detect() honours an explicit api: URL');
-  } else {
-    fail(`breezy.detect() api: → ${JSON.stringify(apiHit)}`);
-  }
-
-  if (breezy.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('breezy.detect() returns null for non-breezy URLs');
-  } else {
-    fail('breezy.detect() should return null for non-breezy URLs');
-  }
-
-  if (breezy.detect({ name: 'X', careers_url: null }) === null && breezy.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('breezy.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('breezy.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: breezy.hr in the PATH (not host) must not be detected.
-  if (breezy.detect({ name: 'Spoof', careers_url: 'https://evil.example/acme.breezy.hr/json' }) === null) {
-    pass('breezy.detect() rejects path-spoofed URLs');
-  } else {
-    fail('breezy.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // parseBreezyResponse — top-level array
-  const sample = [
-    {
-      name: 'Assistant Field Manager',
-      url: 'https://new-incentives.breezy.hr/p/b8e6-assistant-field-manager',
-      published_date: '2026-05-25T14:45:23.799Z',
-      location: { name: 'Niger, Sokoto, NG', city: 'Niger', country: { name: 'NG' }, is_remote: false },
-    },
-    {
-      name: 'Remote Backend Engineer',
-      url: 'https://new-incentives.breezy.hr/p/aa01-backend',
-      location: { city: 'Lagos', state: 'Lagos', country: { name: 'NG' }, is_remote: true },
-    },
-    { name: 'No URL row', location: { name: 'Remote' } },
-    { name: 'Insecure URL', url: 'http://new-incentives.breezy.hr/p/x', location: {} },
-  ];
-  const jobs = parseBreezyResponse(sample, 'New Incentives');
-
-  if (jobs.length === 2) pass('parseBreezyResponse keeps 2 rows (drops missing/non-https url)');
-  else fail(`parseBreezyResponse returned ${jobs.length} rows (expected 2)`);
-
-  if (jobs[0]?.title === 'Assistant Field Manager' && jobs[0]?.company === 'New Incentives' && jobs[0]?.location === 'Niger, Sokoto, NG') {
-    pass('parseBreezyResponse prefers ready-made location.name');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('2026-05-25T14:45:23.799Z')) {
-    pass('parseBreezyResponse parses published_date → postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.location === 'Lagos, Lagos, NG, Remote') {
-    pass('parseBreezyResponse assembles city/state/country and appends Remote');
-  } else {
-    fail(`row 1 location = ${JSON.stringify(jobs[1]?.location)}, expected "Lagos, Lagos, NG, Remote"`);
-  }
-
-  if (jobs[1]?.postedAt === undefined) {
-    pass('parseBreezyResponse omits postedAt when published_date is absent');
-  } else {
-    fail(`row 1 postedAt should be undefined, got ${JSON.stringify(jobs[1]?.postedAt)}`);
-  }
-
-  if (parseBreezyResponse(null, 'X').length === 0 && parseBreezyResponse({}, 'X').length === 0) {
-    pass('non-array payload → empty result (no crash)');
-  } else {
-    fail('non-array payload should yield empty result');
-  }
-
-  // a row already containing "Remote" must not get a duplicate "Remote" suffix
-  const noDup = parseBreezyResponse([{ name: 'R', url: 'https://acme.breezy.hr/p/r', location: { name: 'Remote, EMEA', is_remote: true } }], 'X');
-  if (noDup[0]?.location === 'Remote, EMEA') pass('parseBreezyResponse does not double-append Remote');
-  else fail(`expected "Remote, EMEA", got ${JSON.stringify(noDup[0]?.location)}`);
-
-} catch (e) {
-  fail(`breezy provider tests crashed: ${e.message}`);
 }
 
 // ── 28. OPTIONAL PROFILE PHOTO (opt-in, DACH/European — #264) ────
@@ -5291,912 +4994,6 @@ try {
   }
 } catch (e) {
   fail(`custom instructions test crashed: ${e.message}`);
-}
-
-// ── 30. Provider — comeet ───────────────────────────────────────
-console.log('\n30. Provider — comeet');
-
-try {
-  const comeet = (await import(pathToFileURL(join(ROOT, 'providers/comeet.mjs')).href)).default;
-  const { parseComeetResponse } = await import(pathToFileURL(join(ROOT, 'providers/comeet.mjs')).href);
-
-  if (comeet.id === 'comeet') pass('comeet.id is "comeet"');
-  else fail(`comeet.id is ${JSON.stringify(comeet.id)}`);
-
-  // detect: explicit api: careers-api URL is honoured (and the secret token is
-  // redacted from the informational DetectHit url).
-  const apiUrl = 'https://www.comeet.co/careers-api/2.0/company/30.005/positions?token=ABC123';
-  const apiHit = comeet.detect({ name: 'Spark Hire', api: apiUrl, careers_url: 'https://www.comeet.com/jobs/spark-hire/30.005' });
-  if (apiHit && apiHit.url === 'https://www.comeet.co/careers-api/2.0/company/30.005/positions?token=REDACTED') {
-    pass('comeet.detect() resolves an explicit api: URL and redacts the token');
-  } else {
-    fail(`comeet.detect() api: → ${JSON.stringify(apiHit)}`);
-  }
-
-  // the DetectHit url must not leak the real token (it may be logged)
-  if (apiHit && !apiHit.url.includes('ABC123')) {
-    pass('comeet.detect() does not leak the real token in the DetectHit url');
-  } else {
-    fail(`comeet.detect() leaked the token: ${JSON.stringify(apiHit)}`);
-  }
-
-  // detect: full careers-api URL pasted into careers_url is also accepted
-  const cuHit = comeet.detect({ name: 'X', careers_url: apiUrl });
-  if (cuHit && cuHit.url === 'https://www.comeet.co/careers-api/2.0/company/30.005/positions?token=REDACTED') {
-    pass('comeet.detect() accepts a careers-api URL in careers_url');
-  } else {
-    fail(`comeet.detect() careers_url → ${JSON.stringify(cuHit)}`);
-  }
-
-  // detect: a branded www.comeet.com/jobs page carries no token → not claimed
-  if (comeet.detect({ name: 'X', careers_url: 'https://www.comeet.com/jobs/spark-hire/30.005' }) === null) {
-    pass('comeet.detect() returns null for a branded careers page (no token)');
-  } else {
-    fail('comeet.detect() should not claim a tokenless branded careers page');
-  }
-
-  if (comeet.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('comeet.detect() returns null for non-comeet URLs');
-  } else {
-    fail('comeet.detect() should return null for non-comeet URLs');
-  }
-
-  if (comeet.detect({ name: 'X', careers_url: null }) === null && comeet.detect({ name: 'X', api: 7 }) === null) {
-    pass('comeet.detect() returns null for non-string url fields (null and 7)');
-  } else {
-    fail('comeet.detect() should treat non-string url fields as missing');
-  }
-
-  // SSRF: comeet.co in the PATH (not host) must not be detected.
-  if (comeet.detect({ name: 'Spoof', api: 'https://evil.example/www.comeet.co/careers-api/2.0/company/x/positions' }) === null) {
-    pass('comeet.detect() rejects path-spoofed URLs');
-  } else {
-    fail('comeet.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // SSRF: the wrong comeet host (www.comeet.com, the hosted-page origin) is rejected.
-  if (comeet.detect({ name: 'Spoof', api: 'https://www.comeet.com/careers-api/2.0/company/x/positions?token=y' }) === null) {
-    pass('comeet.detect() pins to www.comeet.co (rejects www.comeet.com)');
-  } else {
-    fail('comeet.detect() must pin to www.comeet.co');
-  }
-
-  // parseComeetResponse — top-level array (real shape, confirmed live)
-  const sample = [
-    {
-      name: 'AI Engineer',
-      url_active_page: 'https://www.comeet.com/jobs/spark-hire/30.005/ai-engineer/F1.B67',
-      url_comeet_hosted_page: 'https://www.comeet.com/jobs/spark-hire/30.005/ai-engineer/F1.B67',
-      time_updated: '2026-06-11T07:49:20Z',
-      location: { name: 'Tel Aviv, Israel', is_remote: true },
-    },
-    {
-      name: 'Backend Engineer',
-      url_comeet_hosted_page: 'https://www.comeet.com/jobs/spark-hire/30.005/backend/AB.C12',
-      location: { name: 'Berlin, Germany', is_remote: false },
-    },
-    { name: 'No URL row', location: { name: 'Remote' } },
-    { name: 'Insecure URL', url_active_page: 'http://www.comeet.com/jobs/x', location: {} },
-  ];
-  const jobs = parseComeetResponse(sample, 'Spark Hire');
-
-  if (jobs.length === 2) pass('parseComeetResponse keeps 2 rows (drops missing/non-https url)');
-  else fail(`parseComeetResponse returned ${jobs.length} rows (expected 2)`);
-
-  if (jobs[0]?.title === 'AI Engineer' && jobs[0]?.company === 'Spark Hire' && jobs[0]?.location === 'Tel Aviv, Israel, Remote') {
-    pass('parseComeetResponse maps name/location.name and appends Remote');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('2026-06-11T07:49:20Z')) {
-    pass('parseComeetResponse parses time_updated → postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.url === 'https://www.comeet.com/jobs/spark-hire/30.005/backend/AB.C12' && jobs[1]?.location === 'Berlin, Germany' && jobs[1]?.postedAt === undefined) {
-    pass('parseComeetResponse falls back to url_comeet_hosted_page and omits absent postedAt');
-  } else {
-    fail(`row 1 = ${JSON.stringify(jobs[1])}`);
-  }
-
-  if (parseComeetResponse(null, 'X').length === 0 && parseComeetResponse({}, 'X').length === 0) {
-    pass('non-array payload → empty result (no crash)');
-  } else {
-    fail('non-array payload should yield empty result');
-  }
-
-  // a location already containing "Remote" must not get a duplicate suffix
-  const noDup = parseComeetResponse([{ name: 'R', url_active_page: 'https://www.comeet.com/jobs/x/r', location: { name: 'Remote, EMEA', is_remote: true } }], 'X');
-  if (noDup[0]?.location === 'Remote, EMEA') pass('parseComeetResponse does not double-append Remote');
-  else fail(`expected "Remote, EMEA", got ${JSON.stringify(noDup[0]?.location)}`);
-
-  // malformed members (null / non-object / whitespace-only name) must neither
-  // throw nor slip through: a row needs a non-empty trimmed title AND a url.
-  const dirty = [
-    null,
-    'not an object',
-    42,
-    { name: '   ', url_active_page: 'https://www.comeet.com/jobs/x/blank' }, // blank title → dropped
-    { name: '  Padded Role  ', url_active_page: 'https://www.comeet.com/jobs/x/p', location: {} }, // trimmed, kept
-  ];
-  const cleaned = parseComeetResponse(dirty, 'X');
-  if (cleaned.length === 1 && cleaned[0].title === 'Padded Role') {
-    pass('parseComeetResponse skips null/non-object/blank-title rows and trims the title');
-  } else {
-    fail(`dirty parse = ${JSON.stringify(cleaned)} (expected 1 row "Padded Role")`);
-  }
-
-} catch (e) {
-  fail(`comeet provider tests crashed: ${e.message}`);
-}
-
-// ── 31. Provider — personio ─────────────────────────────────────
-console.log('\n31. Provider — personio');
-
-try {
-  const personio = (await import(pathToFileURL(join(ROOT, 'providers/personio.mjs')).href)).default;
-  const { parsePersonioXml } = await import(pathToFileURL(join(ROOT, 'providers/personio.mjs')).href);
-
-  if (personio.id === 'personio') pass('personio.id is "personio"');
-  else fail(`personio.id is ${JSON.stringify(personio.id)}`);
-
-  // detect: <slug>.jobs.personio.de careers host → /xml feed
-  const hit = personio.detect({ name: 'Acme', careers_url: 'https://acme.jobs.personio.de/' });
-  if (hit && hit.url === 'https://acme.jobs.personio.de/xml') {
-    pass('personio.detect() resolves <slug>.jobs.personio.de → /xml feed');
-  } else {
-    fail(`personio.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect: the .com TLD variant is also accepted
-  const comHit = personio.detect({ name: 'Acme', careers_url: 'https://acme.jobs.personio.com/jobs' });
-  if (comHit && comHit.url === 'https://acme.jobs.personio.com/xml') {
-    pass('personio.detect() accepts the .com TLD variant');
-  } else {
-    fail(`personio.detect() .com → ${JSON.stringify(comHit)}`);
-  }
-
-  if (personio.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('personio.detect() returns null for non-personio URLs');
-  } else {
-    fail('personio.detect() should return null for non-personio URLs');
-  }
-
-  if (personio.detect({ name: 'X', careers_url: null }) === null && personio.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('personio.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('personio.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: jobs.personio.de in the PATH (not host) must not be detected.
-  if (personio.detect({ name: 'Spoof', careers_url: 'https://evil.example/acme.jobs.personio.de/xml' }) === null) {
-    pass('personio.detect() rejects path-spoofed URLs');
-  } else {
-    fail('personio.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // SSRF: a look-alike host (suffix attack) must be rejected.
-  if (personio.detect({ name: 'Spoof', careers_url: 'https://acme.jobs.personio.de.evil.com/xml' }) === null) {
-    pass('personio.detect() rejects suffix-spoofed look-alike hosts');
-  } else {
-    fail('personio.detect() must reject suffix-spoofed hosts');
-  }
-
-  // parsePersonioXml — the real <workzag-jobs> shape (confirmed live)
-  const HOST = 'acme.jobs.personio.de';
-  const sample = `<?xml version="1.0" encoding="UTF-8"?>
-<workzag-jobs>
-<position>
-  <id>1834171</id>
-  <office>Munich</office>
-  <additionalOffices><office>Berlin</office></additionalOffices>
-  <name>Staff Software Engineer, Data &amp; Platform</name>
-  <createdAt>2024-11-13T14:10:41+00:00</createdAt>
-</position>
-<position>
-  <id>900100</id>
-  <office>Remote</office>
-  <name><![CDATA[Senior Engineer (m/f/d)]]></name>
-  <createdAt>2025-01-02T09:00:00+00:00</createdAt>
-</position>
-<position>
-  <id>777</id>
-  <office>Cologne</office>
-  <name></name>
-</position>
-<position>
-  <id>not-a-number</id>
-  <office>Hamburg</office>
-  <name>Bad ID Role</name>
-</position>
-</workzag-jobs>`;
-  const jobs = parsePersonioXml(sample, 'Acme', HOST);
-
-  if (jobs.length === 2) pass('parsePersonioXml keeps 2 positions (drops empty name + non-numeric id)');
-  else fail(`parsePersonioXml returned ${jobs.length} positions (expected 2)`);
-
-  if (jobs[0]?.title === 'Staff Software Engineer, Data & Platform' && jobs[0]?.company === 'Acme') {
-    pass('parsePersonioXml decodes &amp; in the title');
-  } else {
-    fail(`row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[0]?.url === 'https://acme.jobs.personio.de/job/1834171') {
-    pass('parsePersonioXml builds the job URL from host + numeric id');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.location === 'Munich, Berlin') {
-    pass('parsePersonioXml joins primary + additionalOffices');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}, expected "Munich, Berlin"`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('2024-11-13T14:10:41+00:00')) {
-    pass('parsePersonioXml parses createdAt → postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.title === 'Senior Engineer (m/f/d)') {
-    pass('parsePersonioXml unwraps a CDATA name');
-  } else {
-    fail(`row 1 title = ${JSON.stringify(jobs[1]?.title)}`);
-  }
-
-  if (parsePersonioXml('', 'X', HOST).length === 0 && parsePersonioXml(null, 'X', HOST).length === 0) {
-    pass('empty / non-string feed → empty result (no crash)');
-  } else {
-    fail('empty / non-string feed should yield empty result');
-  }
-
-  // Hardening: <jobDescriptions> carries per-section <name>/<value> pairs whose
-  // nested <name> must NOT be mistaken for the position's own title; numeric
-  // entities decode; an office wrapped in CDATA unwraps.
-  const tricky = `<workzag-jobs><position>
-    <id>42</id>
-    <office><![CDATA[München]]></office>
-    <name>Real Title &#38; More</name>
-    <jobDescriptions>
-      <jobDescription><name>Your tasks</name><value>do things</value></jobDescription>
-    </jobDescriptions>
-    <createdAt>2025-03-04T00:00:00+00:00</createdAt>
-  </position></workzag-jobs>`;
-  const tj = parsePersonioXml(tricky, 'Acme', HOST);
-  if (tj.length === 1 && tj[0].title === 'Real Title & More') {
-    pass('parsePersonioXml ignores nested <jobDescriptions><name> + decodes numeric entity');
-  } else {
-    fail(`tricky title = ${JSON.stringify(tj[0]?.title)} (len ${tj.length})`);
-  }
-  if (tj[0]?.location === 'München') {
-    pass('parsePersonioXml unwraps a CDATA <office>');
-  } else {
-    fail(`tricky location = ${JSON.stringify(tj[0]?.location)}`);
-  }
-
-  // Hardening: a <jobDescriptions> value carrying a literal "</position>" must
-  // not truncate the block split. Stripping descriptions from the whole feed
-  // first keeps both positions intact.
-  const sneaky = `<workzag-jobs><position>
-    <id>1</id><name>First</name>
-    <jobDescriptions><jobDescription><name>About</name><value>uses &lt;/position&gt; literally: </position></value></jobDescription></jobDescriptions>
-  </position><position>
-    <id>2</id><name>Second</name>
-  </position></workzag-jobs>`;
-  const sj2 = parsePersonioXml(sneaky, 'Acme', HOST);
-  if (sj2.length === 2 && sj2[0].title === 'First' && sj2[1].title === 'Second') {
-    pass('parsePersonioXml survives a literal </position> inside <jobDescriptions>');
-  } else {
-    fail(`sneaky parse = ${JSON.stringify(sj2.map(j => j.title))} (len ${sj2.length})`);
-  }
-
-  // fetch() passes redirect:'error' to fetchText (SSRF hardening must not regress)
-  let capturedOpts = null;
-  await personio.fetch(
-    { name: 'Acme', careers_url: 'https://acme.jobs.personio.de/' },
-    { fetchText: async (_url, opts) => { capturedOpts = opts; return '<workzag-jobs></workzag-jobs>'; } },
-  );
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('personio.fetch() passes redirect:"error" to fetchText');
-  } else {
-    fail(`personio.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-} catch (e) {
-  fail(`personio provider tests crashed: ${e.message}`);
-}
-
-// -- 32. Provider - weworkremotely ---------------------------------------
-console.log('\n32. Provider - weworkremotely');
-
-try {
-  const wwrModule = await import(pathToFileURL(join(ROOT, 'providers/weworkremotely.mjs')).href);
-  const weworkremotely = wwrModule.default;
-  const { parseWwrFeed } = wwrModule;
-
-  if (weworkremotely.id === 'weworkremotely') pass('weworkremotely.id is "weworkremotely"');
-  else fail(`weworkremotely.id is ${JSON.stringify(weworkremotely.id)}`);
-
-  const hit = weworkremotely.detect({ name: 'WWR', provider: 'weworkremotely' });
-  if (hit && hit.url === 'https://weworkremotely.com/remote-jobs.rss') {
-    pass('weworkremotely.detect() claims explicit provider config');
-  } else {
-    fail(`weworkremotely.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  if (weworkremotely.detect({ name: 'Remote Board', provider: 'remoteok' }) === null) {
-    pass('weworkremotely.detect() ignores other provider ids');
-  } else {
-    fail('weworkremotely.detect() should only claim provider: weworkremotely');
-  }
-
-  const sample = `<?xml version="1.0" encoding="UTF-8"?>
-<rss version="2.0">
-  <channel>
-    <item>
-      <title><![CDATA[Acme & Co: Staff AI Engineer]]></title>
-      <link>https://weworkremotely.com/remote-jobs/acme-staff-ai-engineer</link>
-      <pubDate>Thu, 13 Nov 2025 14:10:41 +0000</pubDate>
-      <region><![CDATA[Anywhere in the World]]></region>
-      <category>Programming</category>
-    </item>
-    <item>
-      <title>Principal Platform Engineer &amp; Tooling</title>
-      <link>https://weworkremotely.com/remote-jobs/example-platform-engineer</link>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-      <category>DevOps and Sysadmin</category>
-    </item>
-    <item>
-      <title>Missing Link Inc: Dropped Role</title>
-      <pubDate>Fri, 02 Jan 2026 09:00:00 +0000</pubDate>
-      <region>USA Only</region>
-    </item>
-    <item>
-      <title>Bad Link Inc: Dropped Role</title>
-      <link>/remote-jobs/bad-link</link>
-      <region>Europe Only</region>
-    </item>
-    <item>
-      <title>Off Host Inc: Dropped Role</title>
-      <link>https://example.com/remote-jobs/off-host</link>
-      <region>Internal</region>
-    </item>
-  </channel>
-</rss>`;
-  const jobs = parseWwrFeed(sample, 'WWR Board');
-
-  if (jobs.length === 2) pass('parseWwrFeed keeps 2 items (drops missing/relative/off-host links)');
-  else fail(`parseWwrFeed returned ${jobs.length} jobs (expected 2)`);
-
-  if (jobs[0]?.company === 'Acme & Co' && jobs[0]?.title === 'Staff AI Engineer') {
-    pass('parseWwrFeed splits "Company: Role" titles');
-  } else {
-    fail(`row 0 title/company = ${JSON.stringify({ title: jobs[0]?.title, company: jobs[0]?.company })}`);
-  }
-
-  if (jobs[0]?.url === 'https://weworkremotely.com/remote-jobs/acme-staff-ai-engineer') {
-    pass('parseWwrFeed maps <link> to url');
-  } else {
-    fail(`row 0 url = ${JSON.stringify(jobs[0]?.url)}`);
-  }
-
-  if (jobs[0]?.location === 'Anywhere in the World') {
-    pass('parseWwrFeed maps <region> to location');
-  } else {
-    fail(`row 0 location = ${JSON.stringify(jobs[0]?.location)}`);
-  }
-
-  if (jobs[0]?.postedAt === Date.parse('Thu, 13 Nov 2025 14:10:41 +0000')) {
-    pass('parseWwrFeed parses pubDate -> postedAt');
-  } else {
-    fail(`row 0 postedAt = ${JSON.stringify(jobs[0]?.postedAt)}`);
-  }
-
-  if (jobs[1]?.company === 'WWR Board' && jobs[1]?.title === 'Principal Platform Engineer & Tooling') {
-    pass('parseWwrFeed falls back to entry name and decodes entities');
-  } else {
-    fail(`row 1 = ${JSON.stringify(jobs[1])}`);
-  }
-
-  if (jobs[1]?.location === 'DevOps and Sysadmin') {
-    pass('parseWwrFeed falls back to <category> when <region> is absent');
-  } else {
-    fail(`row 1 location = ${JSON.stringify(jobs[1]?.location)}`);
-  }
-
-  if (parseWwrFeed('', 'X').length === 0 && parseWwrFeed(null, 'X').length === 0) {
-    pass('parseWwrFeed empty / non-string feed -> empty result (no crash)');
-  } else {
-    fail('parseWwrFeed empty / non-string feed should yield empty result');
-  }
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await weworkremotely.fetch(
-    { name: 'WWR Board', provider: 'weworkremotely' },
-    { fetchText: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://weworkremotely.com/remote-jobs.rss') {
-    pass('weworkremotely.fetch() requests the pinned RSS feed URL');
-  } else {
-    fail(`weworkremotely.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('weworkremotely.fetch() passes redirect:"error" to fetchText');
-  } else {
-    fail(`weworkremotely.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched[0]?.company === 'Acme & Co' && fetched[0]?.title === 'Staff AI Engineer') {
-    pass('provider: weworkremotely config returns normalized jobs');
-  } else {
-    fail(`weworkremotely.fetch() normalized row = ${JSON.stringify(fetched[0])}`);
-  }
-} catch (e) {
-  fail(`weworkremotely provider tests crashed: ${e.message}`);
-}
-
-// ── 33. Provider — remotive ─────────────────────────────────────
-console.log('\n33. Provider — remotive');
-
-try {
-  const remotiveModule = await import(pathToFileURL(join(ROOT, 'providers/remotive.mjs')).href);
-  const remotive = remotiveModule.default;
-
-  if (remotive.id === 'remotive') pass('remotive.id is "remotive"');
-  else fail(`remotive.id is ${JSON.stringify(remotive.id)}`);
-
-  // Deterministic sample payload — no network. Two valid jobs plus two that must
-  // be dropped by the filter (empty title, non-absolute url).
-  const sample = {
-    jobs: [
-      {
-        title: 'Staff AI Engineer',
-        url: 'https://remotive.com/remote-jobs/acme-staff-ai-engineer',
-        company_name: 'Acme Corp',
-        candidate_required_location: 'Worldwide',
-      },
-      {
-        title: '  Platform Engineer  ',                 // leading/trailing space → trimmed
-        url: '  https://remotive.com/remote-jobs/beta-platform-engineer  ',
-        company_name: '',                               // empty → falls back to entry.name
-        // candidate_required_location omitted → location ''
-      },
-      {
-        title: '',                                       // dropped: empty title
-        url: 'https://remotive.com/remote-jobs/bad-empty-title',
-        company_name: 'Bad Co',
-      },
-      {
-        title: 'Relative URL Role',                      // dropped: non-absolute url
-        url: '/remote-jobs/relative',
-        company_name: 'Rel Co',
-      },
-    ],
-  };
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await remotive.fetch(
-    { name: 'Remotive Board', provider: 'remotive' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://remotive.com/api/remote-jobs')
-    pass('remotive.fetch() requests the board-wide feed URL');
-  else fail(`remotive.fetch() requested ${JSON.stringify(capturedUrl)}`);
-
-  if (capturedOpts && capturedOpts.redirect === 'error')
-    pass('remotive.fetch() passes redirect:"error" to fetchJson (SSRF guard)');
-  else fail(`remotive.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  if (fetched.length === 2)
-    pass('remotive.fetch() keeps 2 valid jobs (drops empty-title + non-absolute-url rows)');
-  else fail(`remotive.fetch() returned ${fetched.length} jobs (expected 2)`);
-
-  // Normalized shape: exactly { title, url, company, location }.
-  if (fetched[0] && Object.keys(fetched[0]).sort().join(',') === 'company,location,title,url')
-    pass('remotive.fetch() returns the normalized { title, url, company, location } shape');
-  else fail(`remotive.fetch() row 0 keys = ${JSON.stringify(fetched[0] && Object.keys(fetched[0]))}`);
-
-  if (fetched[0]?.title === 'Staff AI Engineer'
-      && fetched[0]?.url === 'https://remotive.com/remote-jobs/acme-staff-ai-engineer'
-      && fetched[0]?.company === 'Acme Corp'
-      && fetched[0]?.location === 'Worldwide')
-    pass('remotive.fetch() maps title/url/company_name/candidate_required_location for a full row');
-  else fail(`remotive.fetch() row 0 = ${JSON.stringify(fetched[0])}`);
-
-  if (fetched[1]?.title === 'Platform Engineer'
-      && fetched[1]?.url === 'https://remotive.com/remote-jobs/beta-platform-engineer')
-    pass('remotive.fetch() trims whitespace from title and url');
-  else fail(`remotive.fetch() row 1 title/url = ${JSON.stringify({ title: fetched[1]?.title, url: fetched[1]?.url })}`);
-
-  if (fetched[1]?.company === 'Remotive Board')
-    pass('remotive.fetch() falls back to entry.name when company_name is empty');
-  else fail(`remotive.fetch() row 1 company = ${JSON.stringify(fetched[1]?.company)}`);
-
-  if (fetched[1]?.location === '')
-    pass('remotive.fetch() yields empty location when candidate_required_location is absent');
-  else fail(`remotive.fetch() row 1 location = ${JSON.stringify(fetched[1]?.location)}`);
-
-  // company default when both company_name and entry.name are missing → 'Remotive'.
-  const noName = await remotive.fetch(
-    {},
-    { fetchJson: async () => ({ jobs: [{ title: 'Role', url: 'https://remotive.com/remote-jobs/x' }] }) },
-  );
-  if (noName[0]?.company === 'Remotive')
-    pass('remotive.fetch() defaults company to "Remotive" when company_name and entry.name are both missing');
-  else fail(`remotive.fetch() default company = ${JSON.stringify(noName[0]?.company)}`);
-
-  let badResponseThrew = false;
-  try {
-    await remotive.fetch(
-      { name: 'X', provider: 'remotive' },
-      { fetchJson: async () => ({ wrong: true }) },
-    );
-  } catch (e) {
-    badResponseThrew = /unexpected API response/.test(e.message);
-  }
-  if (badResponseThrew) pass('remotive.fetch() throws on unexpected API response shape');
-  else fail('remotive.fetch() should throw when the jobs array is absent');
-
-} catch (e) {
-  fail(`remotive provider tests crashed: ${e.message}`);
-}
-
-// ── 34. Provider — themuse ─────────────────────────────────────
-console.log('\n34. Provider — themuse');
-
-try {
-  const museModule = await import(pathToFileURL(join(ROOT, 'providers/themuse.mjs')).href);
-  const themuse = museModule.default;
-  const { normalizeMuseJob } = museModule;
-
-  if (themuse.id === 'themuse') pass('themuse.id is "themuse"');
-  else fail(`themuse.id is ${JSON.stringify(themuse.id)}`);
-
-  // normalizeMuseJob — field mapping
-  const job = normalizeMuseJob({
-    name: 'Staff AI Engineer',
-    refs: { landing_page: 'https://www.themuse.com/jobs/acme/staff-ai-engineer' },
-    company: { name: 'Acme Corp' },
-    locations: [{ name: 'Remote' }],
-  });
-  if (job?.title === 'Staff AI Engineer') pass('normalizeMuseJob maps name → title');
-  else fail(`normalizeMuseJob title = ${JSON.stringify(job?.title)}`);
-
-  if (job?.url === 'https://www.themuse.com/jobs/acme/staff-ai-engineer')
-    pass('normalizeMuseJob maps refs.landing_page → url');
-  else fail(`normalizeMuseJob url = ${JSON.stringify(job?.url)}`);
-
-  if (job?.company === 'Acme Corp') pass('normalizeMuseJob maps company.name → company');
-  else fail(`normalizeMuseJob company = ${JSON.stringify(job?.company)}`);
-
-  if (job?.location === 'Remote') pass('normalizeMuseJob maps locations[0].name → location');
-  else fail(`normalizeMuseJob location = ${JSON.stringify(job?.location)}`);
-
-  // Missing/invalid field handling
-  if (normalizeMuseJob({ name: '', refs: { landing_page: 'https://www.themuse.com/jobs/x' }, company: { name: 'X' }, locations: [] }) === null)
-    pass('normalizeMuseJob drops entries with empty title');
-  else fail('normalizeMuseJob should return null for empty title');
-
-  if (normalizeMuseJob({ name: 'Role', refs: { landing_page: '/relative' }, company: { name: 'X' }, locations: [] }) === null)
-    pass('normalizeMuseJob drops entries with a non-absolute landing_page URL');
-  else fail('normalizeMuseJob should return null for a relative URL');
-
-  const noLoc = normalizeMuseJob({
-    name: 'Role', refs: { landing_page: 'https://www.themuse.com/jobs/x' }, company: { name: 'X' }, locations: [],
-  });
-  if (noLoc?.location === '') pass('normalizeMuseJob returns empty location when locations array is empty');
-  else fail(`normalizeMuseJob location for empty locations = ${JSON.stringify(noLoc?.location)}`);
-
-  const noCompany = normalizeMuseJob({
-    name: 'Role', refs: { landing_page: 'https://www.themuse.com/jobs/x' }, company: null, locations: [{ name: 'NYC' }],
-  });
-  if (noCompany?.company === 'The Muse') pass('normalizeMuseJob falls back to "The Muse" when company.name is missing');
-  else fail(`normalizeMuseJob company fallback = ${JSON.stringify(noCompany?.company)}`);
-
-  if (normalizeMuseJob(null) === null && normalizeMuseJob('string') === null)
-    pass('normalizeMuseJob returns null for non-object inputs');
-  else fail('normalizeMuseJob should return null for null/non-object input');
-
-  // fetch() — mock ctx
-  const sampleResults = [
-    {
-      name: 'Staff AI Engineer',
-      refs: { landing_page: 'https://www.themuse.com/jobs/acme/staff-ai-engineer' },
-      company: { name: 'Acme Corp' },
-      locations: [{ name: 'Remote' }],
-    },
-    {
-      name: 'Platform Engineer',
-      refs: { landing_page: 'https://www.themuse.com/jobs/beta/platform-engineer' },
-      company: { name: 'Beta Inc' },
-      locations: [],
-    },
-    {
-      name: '',
-      refs: { landing_page: 'https://www.themuse.com/jobs/bad/role' },
-      company: { name: 'Bad Co' },
-      locations: [],
-    },
-  ];
-
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await themuse.fetch(
-    { name: 'The Muse Board', provider: 'themuse' },
-    {
-      fetchJson: async (url, opts) => {
-        capturedUrl = url; capturedOpts = opts;
-        return { results: sampleResults, page: 0, page_count: 1 };
-      },
-    },
-  );
-
-  if (capturedUrl === 'https://www.themuse.com/api/public/jobs?page=0')
-    pass('themuse.fetch() requests the correct API endpoint');
-  else fail(`themuse.fetch() requested ${JSON.stringify(capturedUrl)}`);
-
-  if (capturedOpts && capturedOpts.redirect === 'error')
-    pass('themuse.fetch() passes redirect:"error" to fetchJson');
-  else fail(`themuse.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-
-  if (fetched.length === 2)
-    pass('themuse.fetch() normalizes 2 valid jobs (drops the empty-title entry)');
-  else fail(`themuse.fetch() returned ${fetched.length} jobs (expected 2)`);
-
-  if (fetched[0]?.title === 'Staff AI Engineer' && fetched[0]?.company === 'Acme Corp')
-    pass('themuse.fetch() returns correct title and company for first result');
-  else fail(`themuse.fetch() row 0 = ${JSON.stringify(fetched[0])}`);
-
-  // Pagination: page_count > 1 causes all pages to be fetched and aggregated.
-  const paginationCalls = [];
-  const page1Result = { name: 'Data Engineer', refs: { landing_page: 'https://www.themuse.com/jobs/acme/de' }, company: { name: 'Acme' }, locations: [] };
-  const paginatedJobs = await themuse.fetch(
-    { name: 'The Muse Board', provider: 'themuse' },
-    {
-      fetchJson: async (url, opts) => {
-        paginationCalls.push(url);
-        const page = parseInt(new URL(url).searchParams.get('page') ?? '0', 10);
-        if (page === 0) return { results: [sampleResults[0]], page: 0, page_count: 2 };
-        return { results: [page1Result], page: 1, page_count: 2 };
-      },
-    },
-  );
-  if (paginationCalls.length === 2 &&
-      paginationCalls[0] === 'https://www.themuse.com/api/public/jobs?page=0' &&
-      paginationCalls[1] === 'https://www.themuse.com/api/public/jobs?page=1')
-    pass('themuse.fetch() iterates all pages when page_count > 1');
-  else fail(`themuse.fetch() pagination calls = ${JSON.stringify(paginationCalls)}`);
-
-  if (paginatedJobs.length === 2 && paginatedJobs[1]?.title === 'Data Engineer')
-    pass('themuse.fetch() aggregates results from all pages');
-  else fail(`themuse.fetch() paginated results = ${JSON.stringify(paginatedJobs.map(j => j.title))}`);
-
-  // page_count cap: a huge value must be clamped to 100, not cause unbounded requests.
-  let cappedCalls = 0;
-  await themuse.fetch(
-    { name: 'X', provider: 'themuse' },
-    { fetchJson: async () => { cappedCalls++; return { results: [], page: 0, page_count: 99999 }; } },
-  );
-  if (cappedCalls === 100) pass('themuse.fetch() clamps page_count to 100 (prevents unbounded requests)');
-  else fail(`themuse.fetch() made ${cappedCalls} requests for page_count=99999 (expected 100)`);
-
-  // Non-integer page_count must be ignored (NaN passes typeof==='number' but not Number.isInteger).
-  let nonIntCalls = 0;
-  await themuse.fetch(
-    { name: 'X', provider: 'themuse' },
-    { fetchJson: async () => { nonIntCalls++; return { results: [], page: 0, page_count: 1.5 }; } },
-  );
-  if (nonIntCalls === 1) pass('themuse.fetch() ignores non-integer page_count (fetches only page 0)');
-  else fail(`themuse.fetch() made ${nonIntCalls} requests for page_count=1.5 (expected 1)`);
-
-  let badResponseThrew = false;
-  try {
-    await themuse.fetch(
-      { name: 'X', provider: 'themuse' },
-      { fetchJson: async () => ({ wrong: true }) },
-    );
-  } catch (e) {
-    badResponseThrew = /unexpected API response/.test(e.message);
-  }
-  if (badResponseThrew) pass('themuse.fetch() throws on unexpected API response shape');
-  else fail('themuse.fetch() should throw when results array is absent');
-
-} catch (e) {
-  fail(`themuse provider tests crashed: ${e.message}`);
-}
-
-// ── 35. Provider — pinpoint ─────────────────────────────────────
-console.log('\n35. Provider — pinpoint');
-
-try {
-  const pinpointModule = await import(pathToFileURL(join(ROOT, 'providers/pinpoint.mjs')).href);
-  const pinpoint = pinpointModule.default;
-  const { parsePinpointResponse } = pinpointModule;
-
-  if (pinpoint.id === 'pinpoint') pass('pinpoint.id is "pinpoint"');
-  else fail(`pinpoint.id is ${JSON.stringify(pinpoint.id)}`);
-
-  // detect(): <slug>.pinpointhq.com careers_url → postings.json endpoint.
-  const hit = pinpoint.detect({ name: 'Pinpoint', careers_url: 'https://workwithus.pinpointhq.com' });
-  if (hit && hit.url === 'https://workwithus.pinpointhq.com/postings.json') {
-    pass('pinpoint.detect() resolves <slug>.pinpointhq.com → postings.json');
-  } else {
-    fail(`pinpoint.detect() returned ${JSON.stringify(hit)}`);
-  }
-
-  // detect() ignores the path/locale on the careers_url — endpoint is host-rooted.
-  const hitWithPath = pinpoint.detect({ name: 'X', careers_url: 'https://acme.pinpointhq.com/en/jobs' });
-  if (hitWithPath && hitWithPath.url === 'https://acme.pinpointhq.com/postings.json') {
-    pass('pinpoint.detect() ignores careers_url path and roots postings.json at the host');
-  } else {
-    fail(`pinpoint.detect() with path returned ${JSON.stringify(hitWithPath)}`);
-  }
-
-  if (pinpoint.detect({ name: 'X', careers_url: 'https://example.com/careers' }) === null) {
-    pass('pinpoint.detect() returns null for non-pinpoint URLs');
-  } else {
-    fail('pinpoint.detect() should return null for non-pinpoint URLs');
-  }
-
-  // careers_url with non-string value → detect() returns null without crashing.
-  if (pinpoint.detect({ name: 'X', careers_url: null }) === null && pinpoint.detect({ name: 'X', careers_url: 7 }) === null) {
-    pass('pinpoint.detect() returns null for non-string careers_url (null and 7)');
-  } else {
-    fail('pinpoint.detect() should treat non-string careers_url as missing');
-  }
-
-  // SSRF: a URL with pinpointhq.com in the PATH (not host) must not be detected.
-  if (pinpoint.detect({ name: 'Spoof', careers_url: 'https://evil.example/acme.pinpointhq.com/foo' }) === null) {
-    pass('pinpoint.detect() rejects path-spoofed URLs');
-  } else {
-    fail('pinpoint.detect() must NOT misdetect path-spoofed URLs');
-  }
-
-  // SSRF: non-https careers_url must not be detected.
-  if (pinpoint.detect({ name: 'Insecure', careers_url: 'http://acme.pinpointhq.com' }) === null) {
-    pass('pinpoint.detect() rejects non-https careers_url');
-  } else {
-    fail('pinpoint.detect() must reject non-https careers_url');
-  }
-
-  // Hostname label must be a valid DNS label — not end (or start) with a hyphen.
-  if (pinpoint.detect({ name: 'Trailing', careers_url: 'https://acme-.pinpointhq.com' }) === null
-      && pinpoint.detect({ name: 'Leading', careers_url: 'https://-acme.pinpointhq.com' }) === null) {
-    pass('pinpoint.detect() rejects tenant labels that start or end with a hyphen');
-  } else {
-    fail('pinpoint.detect() must reject hyphen-edged tenant labels (e.g. acme-.pinpointhq.com)');
-  }
-
-  // A hyphen in the middle of the label is still valid.
-  if (pinpoint.detect({ name: 'Mid', careers_url: 'https://acme-co.pinpointhq.com' })?.url
-      === 'https://acme-co.pinpointhq.com/postings.json') {
-    pass('pinpoint.detect() still accepts internal hyphens (acme-co.pinpointhq.com)');
-  } else {
-    fail('pinpoint.detect() should accept internal hyphens in the tenant label');
-  }
-
-  // parsePinpointResponse — deterministic sample, no network.
-  const sample = {
-    data: [
-      {
-        title: 'Senior Product Manager',
-        url: 'https://workwithus.pinpointhq.com/en/postings/abc-123',
-        location: { id: '283', city: 'London', province: 'London', name: 'Remote' },
-      },
-      {
-        title: '  Backend Engineer  ',                                   // trimmed
-        url: '  https://workwithus.pinpointhq.com/en/postings/def-456  ', // trimmed
-        location: { city: 'Berlin', province: 'Berlin' },                // no name → assembled
-      },
-      {
-        title: 'No Location Role',
-        url: 'https://workwithus.pinpointhq.com/en/postings/ghi-789',
-        // location omitted → ''
-      },
-      { title: '', url: 'https://workwithus.pinpointhq.com/en/postings/jkl' }, // dropped: empty title
-      { title: 'Missing URL Role' },                                          // dropped: no url
-      { title: 'Relative URL Role', url: '/en/postings/mno' },                // dropped: non-absolute
-      { title: 'Insecure URL Role', url: 'http://workwithus.pinpointhq.com/en/postings/pqr' }, // dropped: non-https
-    ],
-  };
-  const jobs = parsePinpointResponse(sample, 'Pinpoint');
-
-  if (jobs.length === 3) pass('parsePinpointResponse keeps 3 valid postings (drops empty-title / no-url / non-absolute / non-https)');
-  else fail(`parsePinpointResponse returned ${jobs.length} postings (expected 3)`);
-
-  // Normalized shape: exactly { title, url, company, location }.
-  if (jobs[0] && Object.keys(jobs[0]).sort().join(',') === 'company,location,title,url') {
-    pass('parsePinpointResponse returns the normalized { title, url, company, location } shape');
-  } else {
-    fail(`parsePinpointResponse row 0 keys = ${JSON.stringify(jobs[0] && Object.keys(jobs[0]))}`);
-  }
-
-  if (jobs[0]?.title === 'Senior Product Manager'
-      && jobs[0]?.url === 'https://workwithus.pinpointhq.com/en/postings/abc-123'
-      && jobs[0]?.company === 'Pinpoint'
-      && jobs[0]?.location === 'Remote') {
-    pass('parsePinpointResponse maps title/url, sets company from entry name, prefers location.name');
-  } else {
-    fail(`parsePinpointResponse row 0 = ${JSON.stringify(jobs[0])}`);
-  }
-
-  if (jobs[1]?.title === 'Backend Engineer'
-      && jobs[1]?.url === 'https://workwithus.pinpointhq.com/en/postings/def-456') {
-    pass('parsePinpointResponse trims whitespace from title and url');
-  } else {
-    fail(`parsePinpointResponse row 1 title/url = ${JSON.stringify({ title: jobs[1]?.title, url: jobs[1]?.url })}`);
-  }
-
-  if (jobs[1]?.location === 'Berlin, Berlin') {
-    pass('parsePinpointResponse assembles location from city/province when name is absent');
-  } else {
-    fail(`parsePinpointResponse row 1 location = ${JSON.stringify(jobs[1]?.location)}, expected "Berlin, Berlin"`);
-  }
-
-  if (jobs[2]?.location === '') {
-    pass('parsePinpointResponse yields empty location when the location object is absent');
-  } else {
-    fail(`parsePinpointResponse row 2 location = ${JSON.stringify(jobs[2]?.location)}`);
-  }
-
-  if (parsePinpointResponse({}, 'X').length === 0) pass('parsePinpointResponse: empty {} → empty result');
-  else fail('parsePinpointResponse: empty {} should yield empty result');
-
-  if (parsePinpointResponse({ data: null }, 'X').length === 0) {
-    pass('parsePinpointResponse: null data → empty result (no crash)');
-  } else {
-    fail('parsePinpointResponse: null data should yield empty result');
-  }
-
-  // fetch(): requests the derived postings.json URL and passes the SSRF guard.
-  let capturedUrl = null;
-  let capturedOpts = null;
-  const fetched = await pinpoint.fetch(
-    { name: 'Pinpoint', careers_url: 'https://workwithus.pinpointhq.com' },
-    { fetchJson: async (url, opts) => { capturedUrl = url; capturedOpts = opts; return sample; } },
-  );
-
-  if (capturedUrl === 'https://workwithus.pinpointhq.com/postings.json') {
-    pass('pinpoint.fetch() requests the derived postings.json URL');
-  } else {
-    fail(`pinpoint.fetch() requested ${JSON.stringify(capturedUrl)}`);
-  }
-
-  if (capturedOpts && capturedOpts.redirect === 'error') {
-    pass('pinpoint.fetch() passes redirect:"error" to fetchJson (SSRF guard)');
-  } else {
-    fail(`pinpoint.fetch() should pass redirect:"error", got: ${JSON.stringify(capturedOpts)}`);
-  }
-
-  if (fetched.length === 3 && fetched[0]?.company === 'Pinpoint') {
-    pass('pinpoint.fetch() returns normalized jobs with company from entry name');
-  } else {
-    fail(`pinpoint.fetch() returned ${fetched.length} jobs, row 0 = ${JSON.stringify(fetched[0])}`);
-  }
-
-  // fetch(): a non-pinpoint careers_url cannot derive an endpoint → throws.
-  let badEntryThrew = false;
-  try {
-    await pinpoint.fetch(
-      { name: 'X', careers_url: 'https://example.com/careers' },
-      { fetchJson: async () => ({ data: [] }) },
-    );
-  } catch (e) {
-    badEntryThrew = /cannot derive API URL/.test(e.message);
-  }
-  if (badEntryThrew) pass('pinpoint.fetch() throws when the careers_url is not a pinpointhq.com host');
-  else fail('pinpoint.fetch() should throw for a non-pinpoint careers_url');
-
-} catch (e) {
-  fail(`pinpoint provider tests crashed: ${e.message}`);
 }
 
 // ── 36. Provider — arbeitnow ────────────────────────────────────
